+++ lib/src/parser/default_parser_algorithm.dart
@@ -52,14 +52,15 @@ class DefaultParserAlgorithm extends ParserAlgorithm {
 
   @override
   String getCardNumber(List<String> inputs) {
-    return inputs.firstWhere(
-      (input) {
-        final cleanValue = input.fixPossibleMisspells();
-        return (cleanValue.length == CardParserConst.cardNumberLength) &&
-            (int.tryParse(cleanValue) ?? -1) != -1;
-      },
-      orElse: () => '',
-    );
+    for (final item in inputs) {
+      final cleanValue = item.fixPossibleMisspells();
+      //debugPrint('getCardNumber: $item => $cleanValue');
+      if (cleanValue.length == CardParserConst.cardNumberLength &&
+          int.tryParse(cleanValue) != null) {
+        return cleanValue;
+      }
+    }
+    return '';
   }
 
   @override