+++ lib/src/utils/scanner_processor.dart
@@ -1,4 +1,5 @@
 import 'dart:async';
+import 'dart:isolate';
 import 'dart:ui';
 
 import 'package:camera/camera.dart';
@@ -6,49 +7,128 @@ import 'package:flutter/foundation.dart';
 import 'package:google_mlkit_text_recognition/google_mlkit_text_recognition.dart';
 import 'package:ml_card_scanner/src/model/card_info.dart';
 import 'package:ml_card_scanner/src/parser/parser_algorithm.dart';
+import 'package:ml_card_scanner/src/utils/image_processor.dart';
+import 'package:ml_card_scanner/src/utils/stream_debouncer.dart';
 
 class ScannerProcessor {
-  final TextRecognizer _recognizer =
-      TextRecognizer(script: TextRecognitionScript.latin);
+
+  static const _kDebugOutputCooldownMillis = 5000;
+  final bool _usePreprocessingFilters;
+  final bool _debugOutputFilteredImage;
+  final TextRecognizer _recognizer = TextRecognizer(
+    script: TextRecognitionScript.latin,
+  );
+  StreamController<Uint8List>? _debugImageStreamController;
+
+  ScannerProcessor({
+    bool usePreprocessingFilters = false,
+    bool debugOutputFilteredImage = false,
+  })  : _usePreprocessingFilters = debugOutputFilteredImage,
+        _debugOutputFilteredImage = debugOutputFilteredImage {
+    if (_debugOutputFilteredImage) {
+      _debugImageStreamController = StreamController<Uint8List>.broadcast();
+    }
+  }
+
+  Stream<Uint8List>? get imageStream =>
+      _debugImageStreamController?.stream.transform(
+        debounceTransformer(
+          const Duration(milliseconds: _kDebugOutputCooldownMillis),
+        ),
+      );
 
   Future<CardInfo?> computeImage(
     ParserAlgorithm parseAlgorithm,
     CameraImage image,
     InputImageRotation rotation,
   ) async {
-    final format = InputImageFormatValue.fromRawValue(
-      image.format.raw,
-    );
+    late InputImage inputImage;
+    if (!_usePreprocessingFilters) {
+      final format = InputImageFormatValue.fromRawValue(
+        image.format.raw,
+      );
+      final plane = image.planes.first;
+      final bytes = image.planes.map((e) => e.bytes).toList();
+      inputImage = await compute<List<Uint8List>, InputImage>(
+        (message) async {
+          final bytesAll = Uint8List.fromList(
+            bytes.fold(
+                <int>[],
+                (List<int> previousValue, element) =>
+                    previousValue..addAll(element)),
+          );
+          final input = InputImage.fromBytes(
+            bytes: bytesAll,
+            metadata: InputImageMetadata(
+              size: Size(image.width.toDouble(), image.height.toDouble()),
+              rotation: rotation,
+              format: format ?? InputImageFormat.yuv420,
+              bytesPerRow: plane.bytesPerRow,
+            ),
+          );
+          return input;
+        },
+        bytes,
+      );
+    } else {
+      final rawFormat = image.format.raw;
+      final rawRotation = rotation.rawValue;
+      final Uint8List bytes = Uint8List.fromList(
+        image.planes.fold(
+          <int>[],
+          (List<int> previousValue, element) =>
+              previousValue..addAll(element.bytes),
+        ),
+      );
+      final width = image.width;
+      final height = image.height;
+      final bytesPerRow = image.planes.first.bytesPerRow;
 
-    final plane = image.planes.first;
-    final bytes = image.planes.map((e) => e.bytes).toList();
+      ReceivePort? receivePort;
+      if (_debugOutputFilteredImage) {
+        receivePort = ReceivePort();
+        receivePort.listen(
+          (message) {
+            if (message is Uint8List) {
+              if (_debugImageStreamController != null &&
+                  !_debugImageStreamController!.isClosed) {
+                _debugImageStreamController?.add(message);
+              }
+            }
+          },
+        );
+      }
+      inputImage = await createInputImageInIsolate(
+        rawBytes: bytes,
+        width: width,
+        height: height,
+        rawRotation: rawRotation,
+        rawFormat: rawFormat,
+        bytesPerRow: bytesPerRow,
+        debugSendPort: receivePort?.sendPort,
+      );
+      receivePort?.close();
+    }
 
-    final inputImage = await compute<List<Uint8List>, InputImage>(
-      (message) async {
-        final bytesAll = Uint8List.fromList(
-          bytes.fold(
-              <int>[],
-              (List<int> previousValue, element) =>
-                  previousValue..addAll(element)),
-        );
-        final input = InputImage.fromBytes(
-          bytes: bytesAll,
-          metadata: InputImageMetadata(
-            size: Size(image.width.toDouble(), image.height.toDouble()),
-            rotation: rotation,
-            format: format ?? InputImageFormat.yuv420,
-            bytesPerRow: plane.bytesPerRow,
-          ),
-        );
-        return input;
-      },
-      bytes,
-    );
     final recognizedText = await _recognizer.processImage(inputImage);
-    return parseAlgorithm.parse(recognizedText);
+
+    /*if (kDebugMode) {
+      debugPrint('\nrecognizedText: ${recognizedText.text}\n');
+      for (var e in recognizedText.blocks) {
+        debugPrint('block: -> ${e.text} ');
+      }
+      debugPrint('\n');
+    }*/
+
+    final parsedCard = await parseAlgorithm.parse(recognizedText);
+    return parsedCard;
   }
 
   void dispose() {
+    if (_debugImageStreamController != null &&
+        !_debugImageStreamController!.isClosed) {
+      _debugImageStreamController?.close();
+    }
     unawaited(_recognizer.close());
   }
 }